Move colormodes.py
