+++ cicd_utils/ridgeplot_examples/_lincoln_weather.py
@@ -7,8 +7,8 @@ if TYPE_CHECKING:
 
     import plotly.graph_objects as go
 
-    from ridgeplot._colormodes import Colormode
-    from ridgeplot._colors import Color, ColorScale
+    from ridgeplot._coloring.colormodes import Colormode
+    from ridgeplot._coloring.colors import Color, ColorScale
 
 
 def main(
