+++ tox.ini
@@ -5,7 +5,6 @@ labels =
 isolated_build = true
 requires =
     tox>=4
-    tox-uv
 
 [testenv]
 description = run the pytest tests under {basepython}
@@ -31,6 +30,7 @@ deps =
     -r requirements/cicd_utils.txt
 allowlist_externals = rm, mv
 commands =
+    rm -rf build/
     rm -f {env:_COV_FILES_RM}
 
     pytest: pytest --no-cov {posargs}
